refactor: flatten package structure with backward compatibility
- Move all modules from modulink/src/ to modulink/ for clean imports
- Add backward compatibility re-exports in modulink/src/ for existing users
- Support both patterns:
  - Preferred: from modulink import Chain, Context, Logging, Timing
  - Deprecated: from modulink.src.chain import Chain (still works)
- Update __init__.py to expose all main classes directly
- Maintain full functionality while improving developer experience

Resolves import path issue where users needed modulink.src.* imports
